[#62] Redesign Media Block and Inline SideBar components.
