+++ src/blocks/inline-sidebar/components/InlineSidebarEditComponent.js
@@ -1,6 +1,7 @@
 import { Component } from "@wordpress/element";
 import { RichText } from "@wordpress/block-editor";
 import { __ } from '@wordpress/i18n';
+import {DescCard} from "../../global-components/DescCard";
 
 export class InlineSidebarEditComponent extends Component {
     constructor() {
@@ -29,36 +30,45 @@ export class InlineSidebarEditComponent extends Component {
 
         return ([
                 <div className={ className }>
-                    <div className={"inline-sidebar"}>
-                        <fieldset>
-                            <legend>{ __("Preface") }</legend>
-                            <RichText
-                                className={"inline-sidebar__input"}
-                                placeholder={ __("Enter a title preface") }
-                                value={ attributes.preface }
-                                onChange={ this.onChangePreface }
-                            />
-                        </fieldset>
+                    <div>
+                        <DescCard
+                            title={"Inline Sidebar"}
+                            hasText={true}
+                            hasImage={false}
+                            hasImages={false}
+                        />
+                        <div className={'contentCard'}>
+                            <fieldset>
+                                <legend>{ __("Preface") }</legend>
+                                <RichText
+                                    className={"contentCard__input"}
+                                    placeholder={ __("Enter a title preface") }
+                                    value={ attributes.preface }
+                                    onChange={ this.onChangePreface }
+                                />
+                            </fieldset>
 
-                        <fieldset>
-                            <legend>{ __("Title") }</legend>
-                            <RichText
-                                className={"inline-sidebar__input"}
-                                placeholder={ __("Enter your block title") }
-                                value={ attributes.title }
-                                onChange={ this.onChangeTitle }
-                            />
-                        </fieldset>
 
-                        <fieldset>
-                            <legend>{ __("Description") }</legend>
-                            <RichText
-                                className={"inline-sidebar__input"}
-                                placeholder={ __("Enter your Description...") }
-                                value={ attributes.description }
-                                onChange={ this.onChangeDescription }
-                            />
-                        </fieldset>
+                            <fieldset>
+                                <legend>{ __("Title") }</legend>
+                                <RichText
+                                    className={"contentCard__input"}
+                                    placeholder={ __("Enter your block title") }
+                                    value={ attributes.title }
+                                    onChange={ this.onChangeTitle }
+                                />
+                            </fieldset>
+
+                            <fieldset>
+                                <legend>{ __("Description") }</legend>
+                                <RichText
+                                    className={"contentCard__input"}
+                                    placeholder={ __("Enter your Description...") }
+                                    value={ attributes.description }
+                                    onChange={ this.onChangeDescription }
+                                />
+                            </fieldset>
+                        </div>
                     </div>
                 </div>
             ]
