+++ src/blocks/inline-sidebar/editor.scss
@@ -1,31 +1,7 @@
 @import "src/common";
 
 .wp-block-alps-gutenberg-blocks-inline-sidebar {
-  border-left: 4px solid $c-blue;
-  padding: 0.3rem;
-  margin-bottom: 80px;
-
-  font-family: "Noto Serif", Georgia, Times, "Times New Roman", serif;
-}
-
-.inline-sidebar {
-  margin-left: 20px;
-
-  &__input {
-    font-family: Montserrat;
-  }
-}
-
-fieldset {
-  display: flex;
-  border-top: 1px solid lightgray;
-  padding: 5px;
-  margin-bottom: 10px;
-}
-
-legend {
-  font-family: Nunito;
-  margin-left: 20px;
-  padding: 0 7px 0 7px;
-  color: $c-blue;
+  background-color: white;
+  border-radius: 8px;
+  box-shadow: 0px 0px 15px rgba(78, 100, 142, 0.15);
 }
