+++ src/blocks/media-block/components/image-editor.js
@@ -1,3 +1,7 @@
+import {Button} from "@wordpress/components";
+import icons from "../../../icons/icons";
+import {__} from "@wordpress/i18n";
+
 const { Component } = wp.element;
 const { withSelect } = wp.data;
 
@@ -25,11 +29,17 @@ class ScaledImage extends Component {
     }
 
     render() {
-        const { url, id } = this.props;
+        const { url, id, onRemove } = this.props;
 
         return (
-            <div>
-                <img src={ url } data-id={ id } onClick={ this.onImageClick } />
+            <div className={'o-image--edit'}>
+                <Button
+                    icon={icons.remoteGalleryItem}
+                    onClick={onRemove}
+                    className={'blocks-gallery-item__remove'}
+                    label={__('Remove Image', 'alps-gutenberg-blocks')}
+                />
+                <img className={"contentCard__image"} src={ url } data-id={ id } />
             </div>
         );
     }
